requirements.txt
Flask==0.9
Flask-Bcrypt==0.5.2
Flask-Cache==0.4.0
Flask-Login==0.1.3
Flask-SQLAlchemy==0.16
Flask-Script==0.3.3
Flask-Testing==0.4
Flask-Uploads==0.1.3
Flask-WTF==0.8
Jinja2==2.6
Pillow==1.7.7
SQLAlchemy==0.7.8
WTForms==1.0.1
Werkzeug==0.8.3
argparse==1.2.1
blinker==1.2
boto==2.5.2
facebook-sdk==0.3.1
flask-urls==0.9.2
gevent==0.13.7
greenlet==0.4.0
gunicorn==0.14.5
httplib2==0.7.4
oauth2==1.5.211
passlib==1.6
psycopg2==2.4.5
py-bcrypt==0.2
simplejson==2.6.0
twill==0.9
wsgiref==0.1.2
